requirements.txt

django==5.0.1
django-ckeditor==6.7.3
django-unfold==0.67.0